Camera: Correct the zoom index calculation.
Correct the formula to calculate the zoom index. The
calculated value was becoming negative randomly. Since
the variable "index" is a int and the calculation involves
a 'int' numerator and a 'float' denominator. Hence casting
the numerator to 'float'.

+++ src/com/android/camera/ui/ZoomController.java
@@ -85,7 +85,7 @@ public class ZoomController extends GLView {
         invalidate();
 
         int index = mRatios.length - 1 - (int)
-                ((position - mSliderTop) /  mValueGap + .5f);
+                ((float)(position - mSliderTop) /  mValueGap + .5f);
         if (index != mIndex || !isMoving) {
             mIndex = index;
             if (mZoomListener != null) {
